Completely move commands into their own modules
also added config loading
