+++ config/example_options.ini
@@ -32,7 +32,7 @@ Spotify_ClientSecret =
 [Cogs]
 # This specify with cogs (group of commands) should be loaded
 # If you have no idea what this does, it's better to leave it as is
-Cogs_Load = help autoplaylist moderate info queue botmanipulate playback dev utility
+Cogs_Load = help autoplaylist moderate info queuemanipulate botmanipulate playback dev utility
 
 [Permissions]
 # This option determines which user has full permissions and control of the bot.
