+++ musicbot/commands/botmanipulate.py
@@ -1,14 +1,43 @@
-async def cmd_disconnect(self, guild):
+import logging
+import discord
+import aiohttp
+from functools import wraps
+
+from ..utils import _get_variable
+from .. import exceptions
+from ..constructs import Response
+
+log = logging.getLogger(__name__)
+
+cog_name = 'botmanipulate'
+
+# @TheerapakG: TODO: move wrappers into one place
+# TODO: Add some sort of `denied` argument for a message to send when someone else tries to use it
+def owner_only(func):
+    @wraps(func)
+    async def wrapper(bot, *args, **kwargs):
+        # Only allow the owner to use these commands
+        orig_msg = _get_variable('message')
+
+        if not orig_msg or orig_msg.author.id == bot.config.owner_id:
+            # noinspection PyCallingNonCallable
+            return await func(bot, *args, **kwargs)
+        else:
+            raise exceptions.PermissionsError("Only the owner can use this command.", expire_in=30)
+
+    return wrapper
+
+async def cmd_disconnect(bot, guild):
     """
     Usage:
         {command_prefix}disconnect
     
     Forces the bot leave the current voice channel.
     """
-    await self.disconnect_voice_client(guild)
+    await bot.disconnect_voice_client(guild)
     return Response("Disconnected from `{0.name}`".format(guild), delete_after=20)
 
-async def cmd_restart(self, channel):
+async def cmd_restart(bot, channel):
     """
     Usage:
         {command_prefix}restart
@@ -17,33 +46,33 @@ async def cmd_restart(self, channel):
     Will not properly load new dependencies or file updates unless fully shutdown
     and restarted.
     """
-    await self.safe_send_message(channel, "\N{WAVING HAND SIGN} Restarting. If you have updated your bot "
+    await bot.safe_send_message(channel, "\N{WAVING HAND SIGN} Restarting. If you have updated your bot "
         "or its dependencies, you need to restart the bot properly, rather than using this command.")
 
-    player = self.get_player_in(channel.guild)
+    player = bot.get_player_in(channel.guild)
     if player and player.is_paused:
         player.resume()
 
-    await self.disconnect_all_voice_clients()
+    await bot.disconnect_all_voice_clients()
     raise exceptions.RestartSignal()
 
-async def cmd_shutdown(self, channel):
+async def cmd_shutdown(bot, channel):
     """
     Usage:
         {command_prefix}shutdown
     
     Disconnects from voice channels and closes the bot process.
     """
-    await self.safe_send_message(channel, "\N{WAVING HAND SIGN}")
+    await bot.safe_send_message(channel, "\N{WAVING HAND SIGN}")
     
-    player = self.get_player_in(channel.guild)
+    player = bot.get_player_in(channel.guild)
     if player and player.is_paused:
         player.resume()
     
-    await self.disconnect_all_voice_clients()
+    await bot.disconnect_all_voice_clients()
     raise exceptions.TerminateSignal()
 
-async def cmd_leaveserver(self, val, leftover_args):
+async def cmd_leaveserver(bot, val, leftover_args):
     """
     Usage:
         {command_prefix}leaveserver <name/ID>
@@ -54,15 +83,15 @@ async def cmd_leaveserver(self, val, leftover_args):
     if leftover_args:
         val = ' '.join([val, *leftover_args])
 
-    t = self.get_guild(val)
+    t = bot.get_guild(val)
     if t is None:
-        t = discord.utils.get(self.guilds, name=val)
+        t = discord.utils.get(bot.guilds, name=val)
         if t is None:
             raise exceptions.CommandError('No guild was found with the ID or name as `{0}`'.format(val))
     await t.leave()
-    return Response('Left the guild: `{0.name}` (Owner: `{0.owner.name}`, ID: `{0.id}`)'.format(t)
+    return Response('Left the guild: `{0.name}` (Owner: `{0.owner.name}`, ID: `{0.id}`)'.format(t))
 
-async def cmd_setnick(self, guild, channel, leftover_args, nick):
+async def cmd_setnick(bot, guild, channel, leftover_args, nick):
     """
     Usage:
         {command_prefix}setnick nick
@@ -83,7 +112,7 @@ async def cmd_setnick(self, guild, channel, leftover_args, nick):
     return Response("Set the bot's nickname to `{0}`".format(nick), delete_after=20)
 
 @owner_only
-async def cmd_setavatar(self, message, url=None):
+async def cmd_setavatar(bot, message, url=None):
     """
     Usage:
         {command_prefix}setavatar [url]
@@ -101,8 +130,8 @@ async def cmd_setavatar(self, message, url=None):
 
     try:
         timeout = aiohttp.ClientTimeout(total=10)
-        async with self.aiosession.get(thing, timeout=timeout) as res:
-            await self.user.edit(avatar=await res.read())
+        async with bot.aiosession.get(thing, timeout=timeout) as res:
+            await bot.user.edit(avatar=await res.read())
 
     except Exception as e:
         raise exceptions.CommandError("Unable to change avatar: {}".format(e), expire_in=20)
@@ -110,7 +139,7 @@ async def cmd_setavatar(self, message, url=None):
     return Response("Changed the bot's avatar.", delete_after=20)
 
 @owner_only
-async def cmd_setname(self, leftover_args, name):
+async def cmd_setname(bot, leftover_args, name):
     """
     Usage:
         {command_prefix}setname name
@@ -122,7 +151,7 @@ async def cmd_setname(self, leftover_args, name):
     name = ' '.join([name, *leftover_args])
 
     try:
-        await self.user.edit(username=name)
+        await bot.user.edit(username=name)
 
     except discord.HTTPException:
         raise exceptions.CommandError(
@@ -135,7 +164,7 @@ async def cmd_setname(self, leftover_args, name):
     return Response("Set the bot's username to **{0}**".format(name), delete_after=20)
 
 @owner_only
-async def cmd_option(self, player, option, value):
+async def cmd_option(bot, player, option, value):
     """
     Usage:
         {command_prefix}option [option] [on/y/enabled/off/n/disabled]
@@ -160,35 +189,35 @@ async def cmd_option(self, player, option, value):
                 'write_current_song']  # these need to match attribute names in the Config class
     if option in ['autoplaylist', 'auto_playlist']:
         if value in bool_y:
-            if self.config.auto_playlist:
-                raise exceptions.CommandError(self.str.get('cmd-option-autoplaylist-enabled', 'The autoplaylist is already enabled!'))
+            if bot.config.auto_playlist:
+                raise exceptions.CommandError(bot.str.get('cmd-option-autoplaylist-enabled', 'The autoplaylist is already enabled!'))
             else:
-                if not self.autoplaylist:
-                    raise exceptions.CommandError(self.str.get('cmd-option-autoplaylist-none', 'There are no entries in the autoplaylist file.'))
-                self.config.auto_playlist = True
-                await self.on_player_finished_playing(player)
+                if not bot.autoplaylist:
+                    raise exceptions.CommandError(bot.str.get('cmd-option-autoplaylist-none', 'There are no entries in the autoplaylist file.'))
+                bot.config.auto_playlist = True
+                await bot.on_player_finished_playing(player)
         elif value in bool_n:
-            if not self.config.auto_playlist:
-                raise exceptions.CommandError(self.str.get('cmd-option-autoplaylist-disabled', 'The autoplaylist is already disabled!'))
+            if not bot.config.auto_playlist:
+                raise exceptions.CommandError(bot.str.get('cmd-option-autoplaylist-disabled', 'The autoplaylist is already disabled!'))
             else:
-                self.config.auto_playlist = False
+                bot.config.auto_playlist = False
         else:
-            raise exceptions.CommandError(self.str.get('cmd-option-invalid-value', 'The value provided was not valid.'))
-        return Response("The autoplaylist is now " + ['disabled', 'enabled'][self.config.auto_playlist] + '.')
+            raise exceptions.CommandError(bot.str.get('cmd-option-invalid-value', 'The value provided was not valid.'))
+        return Response("The autoplaylist is now " + ['disabled', 'enabled'][bot.config.auto_playlist] + '.')
     else:
         is_generic = [o for o in generic if o == option]  # check if it is a generic bool option
         if is_generic and (value in bool_y or value in bool_n):
             name = is_generic[0]
             log.debug('Setting attribute {0}'.format(name))
-            setattr(self.config, name, True if value in bool_y else False)  # this is scary but should work
-            attr = getattr(self.config, name)
+            setattr(bot.config, name, True if value in bool_y else False)  # this is scary but should work
+            attr = getattr(bot.config, name)
             res = "The option {0} is now ".format(option) + ['disabled', 'enabled'][attr] + '.'
             log.warning('Option overriden for this session: {0}'.format(res))
             return Response(res)
         else:
-            raise exceptions.CommandError(self.str.get('cmd-option-invalid-param' ,'The parameters provided were invalid.'))
+            raise exceptions.CommandError(bot.str.get('cmd-option-invalid-param' ,'The parameters provided were invalid.'))
 
-async def cmd_summon(self, channel, guild, author, voice_channel):
+async def cmd_summon(bot, channel, guild, author, voice_channel):
     """
     Usage:
         {command_prefix}summon
@@ -197,9 +226,9 @@ async def cmd_summon(self, channel, guild, author, voice_channel):
     """
 
     if not author.voice:
-        raise exceptions.CommandError(self.str.get('cmd-summon-novc', 'You are not connected to voice. Try joining a voice channel!'))
+        raise exceptions.CommandError(bot.str.get('cmd-summon-novc', 'You are not connected to voice. Try joining a voice channel!'))
 
-    voice_client = self.voice_client_in(guild)
+    voice_client = bot.voice_client_in(guild)
     if voice_client and guild == author.voice.channel.guild:
         await voice_client.move_to(author.voice.channel)
     else:
@@ -209,31 +238,31 @@ async def cmd_summon(self, channel, guild, author, voice_channel):
         if not chperms.connect:
             log.warning("Cannot join channel '{0}', no permission.".format(author.voice.channel.name))
             raise exceptions.CommandError(
-                self.str.get('cmd-summon-noperms-connect', "Cannot join channel `{0}`, no permission to connect.").format(author.voice.channel.name),
+                bot.str.get('cmd-summon-noperms-connect', "Cannot join channel `{0}`, no permission to connect.").format(author.voice.channel.name),
                 expire_in=25
             )
 
         elif not chperms.speak:
             log.warning("Cannot join channel '{0}', no permission to speak.".format(author.voice.channel.name))
             raise exceptions.CommandError(
-                self.str.get('cmd-summon-noperms-speak', "Cannot join channel `{0}`, no permission to speak.").format(author.voice.channel.name),
+                bot.str.get('cmd-summon-noperms-speak', "Cannot join channel `{0}`, no permission to speak.").format(author.voice.channel.name),
                 expire_in=25
             )
 
-        player = await self.get_player(author.voice.channel, create=True, deserialize=self.config.persistent_queue)
+        player = await bot.get_player(author.voice.channel, create=True, deserialize=bot.config.persistent_queue)
 
         if player.is_stopped:
             player.play()
 
-        if self.config.auto_playlist:
-            await self.on_player_finished_playing(player)
+        if bot.config.auto_playlist:
+            await bot.on_player_finished_playing(player)
 
     log.info("Joining {0.guild.name}/{0.name}".format(author.voice.channel))
 
-    return Response(self.str.get('cmd-summon-reply', 'Connected to `{0.name}`').format(author.voice.channel))
+    return Response(bot.str.get('cmd-summon-reply', 'Connected to `{0.name}`').format(author.voice.channel))
 
 @owner_only
-async def cmd_joinserver(self, message, server_link=None):
+async def cmd_joinserver(bot, message, server_link=None):
     """
     Usage:
         {command_prefix}joinserver invite_link
@@ -241,8 +270,8 @@ async def cmd_joinserver(self, message, server_link=None):
     Asks the bot to join a server.  Note: Bot accounts cannot use invite links.
     """
 
-    url = await self.generate_invite_link()
+    url = await bot.generate_invite_link()
     return Response(
-        self.str.get('cmd-joinserver-response', "Click here to add me to a server: \n{}").format(url),
+        bot.str.get('cmd-joinserver-response', "Click here to add me to a server: \n{}").format(url),
         reply=True, delete_after=30
     )
